Move all the base classes of randomizers in a single file
