+++ python/gym_ignition/randomizers/__init__.py
@@ -2,7 +2,7 @@
 # This software may be modified and distributed under the terms of the
 # GNU Lesser General Public License v2.1 or any later version.
 
-from . import base
+from . import abc
 from . import model
 from . import physics
