--- python/gym_ignition/randomizers/base/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (C) 2020 Istituto Italiano di Tecnologia (IIT). All rights reserved.
-# This software may be modified and distributed under the terms of the
-# GNU Lesser General Public License v2.1 or any later version.
-
-from . import task
-from . import model
-from . import physics
